+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * `-flusher.wal-dir` for the WAL directory to recover from.
   * `-flusher.concurrent-flushes` for number of concurrent flushes.
   * `-flusher.flush-op-timeout` is duration after which a flush should timeout.
+* [ENHANCEMENT] Experimental TSDB: Add support for local `filesystem` backend. #2245
 
 ## 0.7.0-rc.0 / 2020-03-09